- making products unique

+++ database/migrations/2024_11_30_181002_create_products_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->string('name');
             $table->string('image')->nullable();
             $table->string('description');
-            $table->integer('price');
-            $table->integer('available_quantity');
+            $table->unsignedBigInteger('price');
+            $table->unsignedInteger('available_quantity');
             $table->timestamps();
             $table->foreignIdFor(Market::class);
         });
